Add support for Phraseapp branches
Background
==========
As of now, it is not possible to configue the Phraseapp branches we want to download and thus we are stuck with the default, master branch. This was also due to the fact that Phraseapp Java Client had no branching support prior to version 1.0.7.

Goal
====
Since version 1.0.7, the Phraseapp Java Client supports Phraseapp branches. The goal is to start using version 1.0.7 of Phraseapp Java Client and allow the configuration of the branch(es) to be downloaded.